마이페이지 dp조회 추가

+++ src/main/java/com/walkd/dmzing/domain/User.java
@@ -26,7 +26,7 @@ public class User {
     private String authority;
     private String phoneNumber;
 
-    private int dmzPoint = 500;
+    private Long dmzPoint = 500L;
 
 
     @Builder
